Add files via upload
The Code used in the Thesis Paper named "An Optimized Model Using AHP-VIKOR
in the Web-Service Based E-Learning" (author: Avirup Saha) for the implementation of AHP and VIKOR methodology. Here the AHP code is used for the decision-making using the pairwise matrix and using the concept of Best and Worst value the VIKOR method is applied for ranking.